Split Browser runtime dispatch helpers
