+++ extensions/browser/src/browser-tool.ts
@@ -4,98 +4,29 @@
  * Builds the model-facing browser tool, chooses sandbox/host/node routing, and
  * maps high-level actions onto browser control client calls.
  */
-/* oxlint-disable max-lines -- The model-facing browser dispatcher keeps one shared route/approval pipeline; action implementations remain split into focused modules. */
 import type { AgentToolResult } from "openclaw/plugin-sdk/agent-core";
-import { asNullableRecord, isRecord } from "openclaw/plugin-sdk/string-coerce-runtime";
+import { asNullableRecord } from "openclaw/plugin-sdk/string-coerce-runtime";
 import { truncateUtf16Safe } from "openclaw/plugin-sdk/text-utility-runtime";
-import {
-  createBrowserNodeProxyRequest,
-  createBrowserNodeSessionTabRoute,
-  type BrowserOwnedGatewayRequest,
-} from "./browser-node-proxy.js";
-import {
-  applyBrowserTabToolBinding,
-  applyBrowserTabToolBindingToParams,
-  parseBrowserTabToolBinding,
-  type BrowserTabToolBinding,
-} from "./browser-tool-binding.js";
+import { parseBrowserTabToolBinding } from "./browser-tool-binding.js";
 import { describeBrowserTool } from "./browser-tool-description.js";
+import { createBrowserToolExecutor, type BrowserToolOptions } from "./browser-tool-executor.js";
+import type { AnyAgentTool } from "./browser-tool.runtime.js";
 import {
-  createBrowserToolSessionTabs,
-  stripBrowserOpenInternalMetadata,
-} from "./browser-tool-session-tabs.js";
-import {
-  executeActAction,
-  executeConsoleAction,
-  executeDownloadAction,
-  executeEmulateAction,
-  executeRequestsAction,
-  executeErrorsAction,
-  executeTextAction,
-  executeTabsAction,
-  formatBrowserExternalToolResult,
-} from "./browser-tool.actions.js";
-import { executeBrowserLifecycleAction } from "./browser-tool.lifecycle.js";
-import {
-  resolveBrowserBaseUrl,
-  resolveBrowserToolNodeTarget,
-  resolveBrowserToolTimeoutMs,
-  type BrowserNodeTarget,
-} from "./browser-tool.routing.js";
-import {
-  type AnyAgentTool,
   BrowserToolOutputSchema,
   createBrowserToolSchema,
-  resolveBrowserToolCapabilities,
-  type BrowserToolCapabilities,
-  browserAct,
-  browserArmDialog,
-  browserArmFileChooser,
-  browserCloseTab,
-  browserFocusTab,
-  browserNavigate,
-  browserOpenTab,
-  browserPdfSave,
-  callGatewayTool,
-  getRuntimeConfig,
   getBrowserProfileCapabilities,
-  jsonResult,
-  normalizeOptionalString,
-  readPositiveIntegerParam,
+  getRuntimeConfig,
+  resolveBrowserConfig,
   readStringParam,
   readStringValue,
-  resolveBrowserConfig,
-  resolveExistingUploadPaths,
+  resolveBrowserToolCapabilities,
   resolveProfile,
-  touchSessionBrowserTab,
-  trackSessionBrowserTab,
-  untrackSessionBrowserTab,
 } from "./browser-tool.runtime.js";
-import {
-  executeScreenshotAction,
-  type BrowserScreenshotOptions,
-} from "./browser-tool.screenshot.js";
-import { appendNavigatedPageState, executeSnapshotAction } from "./browser-tool.snapshot.js";
-import { resolveBrowserNavigationTimeoutMs } from "./browser/act-policy.js";
-import {
-  createBrowserStewardGatewayApprovalClaim,
-  finalizeBrowserStewardRuntimeParams,
-  getBrowserStewardRuntimeApprovalBinding,
-  isBrowserStewardRuntimeApproved,
-  matchesBrowserStewardRuntimeApprovalBindingAtExecution,
-  prepareBrowserStewardRuntimeParams,
-  resolveBrowserStewardRuntimePolicyParams,
-  resolveBrowserStewardRuntimeApprovedParams,
-  resolveBrowserStewardRuntimeApprovalBinding,
-  type BrowserStewardRuntimeApprovalAuthority,
-  type BrowserStewardRuntimeApprovalBinding,
-} from "./browser/browser-steward-approval.js";
-import {
-  assertBrowserStewardRuntimeAllowed,
-  shouldApplyBrowserStewardRuntimeGuard,
-  type BrowserStewardRuntimeDecision,
-} from "./browser/browser-steward-runtime-guard.js";
-import { parseBrowserNavigationUrl } from "./browser/navigation-guard.js";
+import { finalizeBrowserStewardRuntimeParams } from "./browser/browser-steward-approval.js";
+import { prepareBrowserStewardToolParams } from "./browser/browser-steward-tool-params.js";
+import { readActRequestParam } from "./browser/browser-tool-input.js";
+
+export { prepareBrowserStewardToolParams };
 
 function withBrowserTabDetails(
   result: AgentToolResult<unknown>,
@@ -136,282 +67,8 @@ function withBrowserTabDetails(
   }
 }
 
-function readOptionalTargetAndTimeout(params: Record<string, unknown>) {
-  const targetId = normalizeOptionalString(params.targetId);
-  const timeoutMs = readPositiveIntegerParam(params, "timeoutMs", {
-    message: "timeoutMs must be a positive integer.",
-  });
-  return { targetId, timeoutMs };
-}
-
-function readTargetUrlParam(params: Record<string, unknown>) {
-  const targetUrl =
-    readStringParam(params, "targetUrl") ??
-    readStringParam(params, "url", { required: true, label: "targetUrl" });
-  parseBrowserNavigationUrl(targetUrl);
-  return targetUrl;
-}
-
-const LEGACY_BROWSER_ACT_REQUEST_KEYS = [
-  "kind",
-  "actions",
-  "stopOnError",
-  "targetId",
-  "ref",
-  "doubleClick",
-  "button",
-  "modifiers",
-  "x",
-  "y",
-  "text",
-  "submit",
-  "slowly",
-  "key",
-  "delayMs",
-  "startRef",
-  "endRef",
-  "values",
-  "fields",
-  "width",
-  "height",
-  "timeMs",
-  "textGone",
-  "selector",
-  "url",
-  "loadState",
-  "fn",
-  "timeoutMs",
-] as const;
-
-const LEGACY_BROWSER_ACT_SHARED_REQUEST_KEYS = new Set<
-  (typeof LEGACY_BROWSER_ACT_REQUEST_KEYS)[number]
->(["targetId"]);
-
-function readActRequestParam(params: Record<string, unknown>) {
-  const requestParam = params.request;
-  if (requestParam && typeof requestParam === "object") {
-    const request = { ...(requestParam as Record<string, unknown>) };
-    const hasMismatchedKind =
-      typeof request.kind === "string" &&
-      typeof params.kind === "string" &&
-      request.kind !== params.kind;
-    for (const key of LEGACY_BROWSER_ACT_REQUEST_KEYS) {
-      if (Object.hasOwn(request, key) || !Object.hasOwn(params, key)) {
-        continue;
-      }
-      // Flattened act fields are legacy shape repair. Only the tab scope is
-      // safe across kind mismatches; action-specific fields can corrupt the
-      // explicit nested request.
-      if (hasMismatchedKind && !LEGACY_BROWSER_ACT_SHARED_REQUEST_KEYS.has(key)) {
-        continue;
-      }
-      request[key] = params[key];
-    }
-    return request as Parameters<typeof browserAct>[1];
-  }
-
-  const kind = readStringParam(params, "kind");
-  if (!kind) {
-    return undefined;
-  }
-
-  const request: Record<string, unknown> = {};
-  for (const key of LEGACY_BROWSER_ACT_REQUEST_KEYS) {
-    if (!Object.hasOwn(params, key)) {
-      continue;
-    }
-    request[key] = params[key];
-  }
-  return request as Parameters<typeof browserAct>[1];
-}
-
-async function acquireBrowserNodeSessionLease(
-  nodeId: string,
-  signal?: AbortSignal,
-  existingLease?: string,
-): Promise<string> {
-  const response = await callGatewayTool(
-    "browser.request",
-    { timeoutMs: 10_000 },
-    {
-      routeOnly: true,
-      nodeId,
-      ...(existingLease
-        ? {
-            browserNodeSessionLease: existingLease,
-            renewBrowserNodeSessionLease: true,
-          }
-        : {}),
-    },
-    { scopes: ["operator.admin"], ...(signal ? { signal } : {}) },
-  );
-  const envelope = response;
-  let payload: unknown = envelope?.payload ?? response;
-  if (payload === undefined && typeof envelope?.payloadJSON === "string") {
-    try {
-      payload = JSON.parse(envelope.payloadJSON);
-    } catch {
-      payload = undefined;
-    }
-  }
-  const lease = isRecord(payload) ? payload.browserNodeSessionLease : undefined;
-  if (typeof lease !== "string" || !lease.trim()) {
-    throw new Error("browser node route lease unavailable");
-  }
-  return lease.trim();
-}
-
-async function resolveBrowserStewardToolBinding(params: {
-  input: Record<string, unknown>;
-  agentSessionKey?: string;
-  agentId?: string;
-  sandboxBridgeUrl?: string;
-  allowHostControl?: boolean;
-  signal?: AbortSignal;
-}): Promise<BrowserStewardRuntimeApprovalBinding | undefined> {
-  if (
-    !shouldApplyBrowserStewardRuntimeGuard({
-      sessionKey: params.agentSessionKey,
-      agentId: params.agentId,
-    })
-  ) {
-    return undefined;
-  }
-  const action = typeof params.input.action === "string" ? params.input.action.trim() : "";
-  const requestedTargetValue =
-    typeof params.input.target === "string" ? params.input.target.trim().toLowerCase() : "";
-  const requestedTarget =
-    requestedTargetValue === "sandbox"
-      ? "sandbox"
-      : requestedTargetValue === "host"
-        ? "host"
-        : requestedTargetValue === "node"
-          ? "node"
-          : "";
-  const requestedNode =
-    typeof params.input.node === "string" ? params.input.node.trim() : undefined;
-  const requestedProfile =
-    typeof params.input.profile === "string" ? params.input.profile.trim() : undefined;
-  const runtimeConfig = getRuntimeConfig();
-  const resolvedBrowser = resolveBrowserConfig(runtimeConfig.browser, runtimeConfig);
-  const resolvedProfile = resolveProfile(
-    resolvedBrowser,
-    requestedProfile ?? resolvedBrowser.defaultProfile,
-  );
-  const isUserBrowserProfile = Boolean(
-    resolvedProfile && getBrowserProfileCapabilities(resolvedProfile).usesChromeMcp,
-  );
-  let target: "sandbox" | "host" | "node" | "" = requestedTarget;
-  if (action === "importprofile") {
-    target = "host";
-  }
-  let nodeTarget: BrowserNodeTarget | null = null;
-  if (target !== "host" && target !== "sandbox") {
-    try {
-      nodeTarget = await resolveBrowserToolNodeTarget({
-        requestedNode,
-        target: target || undefined,
-        sandboxBridgeUrl: params.sandboxBridgeUrl,
-        allowHostControl: params.allowHostControl,
-        signal: params.signal,
-      });
-    } catch (error) {
-      params.signal?.throwIfAborted();
-      const configuredNode = runtimeConfig.gateway?.nodes?.browser?.node?.trim();
-      if (!(isUserBrowserProfile && !target && !requestedNode && !configuredNode)) {
-        throw error;
-      }
-    }
-  }
-  if (isUserBrowserProfile && !target && !requestedNode && !nodeTarget) {
-    target = "host";
-  }
-  const browserNodeSessionLease = nodeTarget
-    ? await acquireBrowserNodeSessionLease(nodeTarget.nodeId, params.signal)
-    : undefined;
-  const backendKind = nodeTarget
-    ? "node"
-    : target || (params.sandboxBridgeUrl ? "sandbox" : "host");
-  const origin =
-    backendKind === "node"
-      ? undefined
-      : resolveBrowserBaseUrl({
-          target: backendKind === "sandbox" ? "sandbox" : "host",
-          sandboxBridgeUrl: params.sandboxBridgeUrl,
-          allowHostControl: params.allowHostControl,
-        });
-  const bindingInput = {
-    ...params.input,
-    target: backendKind,
-    ...(nodeTarget ? { node: nodeTarget.nodeId, targetRef: nodeTarget.nodeId } : {}),
-    ...(origin ? { origin } : {}),
-    ...(browserNodeSessionLease ? { browserNodeSessionLease } : {}),
-    ...(backendKind === "node"
-      ? action === "profiles"
-        ? {}
-        : { profile: requestedProfile ?? resolvedBrowser.defaultProfile }
-      : { profile: requestedProfile ?? resolvedBrowser.defaultProfile }),
-  };
-  return resolveBrowserStewardRuntimeApprovalBinding(bindingInput);
-}
-
-export async function prepareBrowserStewardToolParams(params: {
-  input: unknown;
-  agentSessionKey?: string;
-  agentId?: string;
-  sandboxBridgeUrl?: string;
-  allowHostControl?: boolean;
-  runToolBinding?: BrowserTabToolBinding;
-  approvalAuthority?: BrowserStewardRuntimeApprovalAuthority;
-  signal?: AbortSignal;
-}): Promise<unknown> {
-  const input =
-    params.runToolBinding && params.input && typeof params.input === "object"
-      ? applyBrowserTabToolBindingToParams(params.input, params.runToolBinding)
-      : params.input;
-  if (
-    !shouldApplyBrowserStewardRuntimeGuard({
-      sessionKey: params.agentSessionKey,
-      agentId: params.agentId,
-    })
-  ) {
-    return input;
-  }
-  const binding = isRecord(input)
-    ? await resolveBrowserStewardToolBinding({
-        input,
-        agentSessionKey: params.agentSessionKey,
-        agentId: params.agentId,
-        sandboxBridgeUrl: params.sandboxBridgeUrl,
-        allowHostControl: params.allowHostControl,
-        signal: params.signal,
-      })
-    : undefined;
-  return prepareBrowserStewardRuntimeParams(input, binding, params.approvalAuthority);
-}
-
-function readToolTimeoutMs(params: Record<string, unknown>) {
-  return readPositiveIntegerParam(params, "timeoutMs", {
-    message: "timeoutMs must be a positive integer.",
-  });
-}
-
 /** Create the Browser tool exposed to agents. */
-export function createBrowserTool(
-  opts?: BrowserScreenshotOptions & {
-    sandboxBridgeUrl?: string;
-    allowHostControl?: boolean;
-    agentSessionKey?: string;
-    /** Trusted Gateway owner identity; never read from model arguments. */
-    senderIsOwner?: boolean;
-    runToolBinding?: unknown;
-    /** Browser-owned approval authority; never exposed to model-visible tool arguments. */
-    approvalAuthority?: BrowserStewardRuntimeApprovalAuthority;
-    /** Browser-owned lifecycle path for retained node-tab cleanup. */
-    browserOwnedGatewayRequest?: BrowserOwnedGatewayRequest;
-    toolCapabilities?: BrowserToolCapabilities;
-  },
-): AnyAgentTool {
+export function createBrowserTool(opts?: BrowserToolOptions): AnyAgentTool {
   const bindingResult =
     opts?.runToolBinding === undefined
       ? undefined
@@ -461,541 +118,11 @@ export function createBrowserTool(
       }),
     finalizeBeforeToolCallParams: (params, preparedParams) =>
       finalizeBrowserStewardRuntimeParams(params, preparedParams, opts?.approvalAuthority),
-    execute: async (_toolCallId, args, signal) => {
-      const publicParams = bindingResult?.ok
-        ? applyBrowserTabToolBinding(args as Record<string, unknown>, bindingResult.binding)
-        : (args as Record<string, unknown>);
-      const approved = isBrowserStewardRuntimeApproved(publicParams, opts?.approvalAuthority);
-      const browserStewardGatewayApproval = approved
-        ? (request: Parameters<typeof createBrowserStewardGatewayApprovalClaim>[0]) =>
-            createBrowserStewardGatewayApprovalClaim(request)
-        : undefined;
-      const approvedBinding = approved
-        ? getBrowserStewardRuntimeApprovalBinding(publicParams, opts?.approvalAuthority)
-        : undefined;
-      const appliesBrowserStewardRuntimeGuard = shouldApplyBrowserStewardRuntimeGuard({
-        sessionKey: opts?.agentSessionKey,
-        agentId: opts?.agentId,
-      });
-      let browserStewardRuntimeDecision: BrowserStewardRuntimeDecision | undefined;
-      if (appliesBrowserStewardRuntimeGuard) {
-        const policyParams = resolveBrowserStewardRuntimePolicyParams(
-          publicParams,
-          opts?.approvalAuthority,
-        );
-        browserStewardRuntimeDecision = assertBrowserStewardRuntimeAllowed({
-          action: readStringParam(publicParams, "action", { required: true }),
-          profile: readStringParam(publicParams, "profile"),
-          agentSessionKey: opts?.agentSessionKey,
-          agentId: opts?.agentId,
-          approved,
-          request: policyParams.request ?? policyParams,
-        });
-      }
-      const params = approved
-        ? resolveBrowserStewardRuntimeApprovedParams(publicParams, opts?.approvalAuthority)
-        : publicParams;
-      const action = readStringParam(params, "action", { required: true });
-      if (!capabilities.actions.some((candidate) => candidate === action)) {
-        throw new Error(
-          `browser action ${JSON.stringify(action)} is unavailable for this run; use an available action such as snapshot, or select a managed browser profile in an unbound run.`,
-        );
-      }
-      const requestedProfile = readStringParam(params, "profile");
-      const requestedNode = readStringParam(params, "node");
-      const requestedTimeoutMs = readToolTimeoutMs(params);
-      let target = readStringParam(params, "target") as "sandbox" | "host" | "node" | undefined;
-      const runtimeConfig = getRuntimeConfig();
-      const resolvedBrowser = resolveBrowserConfig(runtimeConfig.browser, runtimeConfig);
-      const effectiveProfile = requestedProfile ?? resolvedBrowser.defaultProfile;
-      const resolvedProfile = resolveProfile(resolvedBrowser, effectiveProfile);
-      const profileCapabilities = resolvedProfile
-        ? getBrowserProfileCapabilities(resolvedProfile)
-        : undefined;
-      let profile = profileCapabilities?.usesChromeMcp ? effectiveProfile : requestedProfile;
-      const configuredNode = runtimeConfig.gateway?.nodes?.browser?.node?.trim();
-
-      if (requestedNode && target && target !== "node") {
-        throw new Error('node is only supported with target="node".');
-      }
-
-      // System-profile import reads the local macOS Keychain and Chrome profile,
-      // so it can only run on the host. Pin it before target/node resolution so a
-      // sandbox default or auto-selected browser node never receives the request.
-      if (action === "importprofile") {
-        if (target === "sandbox" || target === "node" || requestedNode) {
-          throw new Error(
-            'system profile import must run on the host; omit target or use target="host".',
-          );
-        }
-        target = "host";
-      }
-      // existing-session profiles can attach through the selected host or browser node,
-      // but they must never fall back into the sandbox browser.
-      const isUserBrowserProfile = profileCapabilities?.usesChromeMcp === true;
-      if (isUserBrowserProfile) {
-        if (target === "sandbox") {
-          throw new Error(
-            `profile="${profile}" cannot use the sandbox browser; use target="host" or omit target.`,
-          );
-        }
-      }
-
-      let nodeTarget: BrowserNodeTarget | null = null;
-      try {
-        nodeTarget = await resolveBrowserToolNodeTarget({
-          requestedNode: requestedNode ?? undefined,
-          target,
-          sandboxBridgeUrl: opts?.sandboxBridgeUrl,
-          allowHostControl: opts?.allowHostControl,
-          signal,
-        });
-      } catch (error) {
-        signal?.throwIfAborted();
-        // Keep the logged-in user browser usable on the host when auto-discovery
-        // of browser nodes fails transiently. Explicit node requests still fail.
-        if (!(isUserBrowserProfile && !target && !requestedNode && !configuredNode)) {
-          throw error;
-        }
-      }
-      if (isUserBrowserProfile && !target && !requestedNode && !nodeTarget) {
-        target = "host";
-      }
-
-      const resolvedTarget = target === "node" ? undefined : target;
-      const baseUrl = nodeTarget
-        ? undefined
-        : resolveBrowserBaseUrl({
-            target: resolvedTarget,
-            sandboxBridgeUrl: opts?.sandboxBridgeUrl,
-            allowHostControl: opts?.allowHostControl,
-          });
-
-      if (approved) {
-        if (nodeTarget && approvedBinding?.browserNodeSessionLease) {
-          await acquireBrowserNodeSessionLease(
-            nodeTarget.nodeId,
-            signal,
-            approvedBinding.browserNodeSessionLease,
-          );
-        }
-        const actualBinding = resolveBrowserStewardRuntimeApprovalBinding({
-          ...publicParams,
-          target: nodeTarget ? "node" : (target ?? (opts?.sandboxBridgeUrl ? "sandbox" : "host")),
-          ...(nodeTarget ? { node: nodeTarget.nodeId, targetRef: nodeTarget.nodeId } : {}),
-          ...(baseUrl ? { origin: baseUrl } : {}),
-          ...(approvedBinding?.browserNodeSessionLease
-            ? { browserNodeSessionLease: approvedBinding.browserNodeSessionLease }
-            : {}),
-          ...(nodeTarget
-            ? action === "profiles"
-              ? {}
-              : { profile: requestedProfile ?? effectiveProfile }
-            : { profile: effectiveProfile }),
-        });
-        if (
-          !approvedBinding ||
-          !matchesBrowserStewardRuntimeApprovalBindingAtExecution(approvedBinding, actualBinding)
-        ) {
-          throw new Error("Browser Steward approval route changed; request approval again");
-        }
-      }
-
-      const allowAutomaticHostFallback = Boolean(
-        nodeTarget &&
-        !approved &&
-        !target &&
-        !requestedNode &&
-        !configuredNode &&
-        opts?.allowHostControl !== false,
-      );
-      const proxyRequest = nodeTarget
-        ? createBrowserNodeProxyRequest({
-            nodeTarget,
-            allowAutomaticHostFallback,
-            agentSessionKey: opts?.agentSessionKey,
-            agentId: opts?.agentId,
-            browserNodeSessionLease: approvedBinding?.browserNodeSessionLease,
-            browserStewardGatewayApproval,
-            signal,
-          })
-        : null;
-      if (proxyRequest) {
-        // Normal node requests resolve omissions on the node; approved Steward
-        // requests pin the profile before approval so the effect cannot drift.
-        profile = browserStewardRuntimeDecision
-          ? action === "profiles"
-            ? undefined
-            : (requestedProfile ?? effectiveProfile)
-          : requestedProfile;
-      }
-      if (
-        !proxyRequest &&
-        isUserBrowserProfile &&
-        ["requests", "errors", "text", "emulate"].includes(action)
-      ) {
-        throw new Error(
-          `action=${action} is not supported for existing-session profiles; use action=snapshot to inspect this page, or select a managed browser profile for ${action}.`,
-        );
-      }
-      const nodeRoute = nodeTarget
-        ? createBrowserNodeSessionTabRoute({
-            nodeTarget,
-            agentSessionKey: opts?.agentSessionKey,
-            agentId: opts?.agentId,
-            browserNodeSessionLease: approvedBinding?.browserNodeSessionLease,
-            browserStewardGatewayApproval,
-            browserOwnedGatewayRequest: opts?.browserOwnedGatewayRequest,
-          })
-        : undefined;
-      const toolTimeoutMs = resolveBrowserToolTimeoutMs({
-        requestedTimeoutMs,
-        action,
-        isUserBrowserProfile,
-        resolvedBrowser,
-      });
-      const sessionTabs = createBrowserToolSessionTabs({
-        sessionKey: opts?.agentSessionKey,
-        requestedProfile: profile,
-        defaultProfile: resolvedBrowser.defaultProfile,
-        baseUrl,
-        nodeRoute,
-        routeProfile: () => {
-          const route = proxyRequest?.route();
-          return route?.status === "resolved" ? route.profile : undefined;
-        },
-        isHostFallbackActive: proxyRequest?.isHostFallbackActive,
-        registry: { touchSessionBrowserTab, trackSessionBrowserTab, untrackSessionBrowserTab },
-        ...(browserStewardRuntimeDecision ? { browserStewardRuntimeDecision } : {}),
-      });
-      const executeTrackedTabRequest = async (
-        path: string,
-        body: Record<string, unknown>,
-        runLocal: () => Promise<unknown>,
-      ) => {
-        const result = proxyRequest
-          ? await proxyRequest({ method: "POST", path, profile, body })
-          : await runLocal();
-        sessionTabs.touch(
-          readStringValue((result as { targetId?: unknown }).targetId) ??
-            readStringValue(body.targetId),
-        );
-        return jsonResult(result);
-      };
-
-      switch (action) {
-        case "doctor":
-        case "status":
-        case "start":
-        case "stop":
-        case "profiles":
-        case "importprofile":
-          return await executeBrowserLifecycleAction({
-            action,
-            input: params,
-            baseUrl,
-            profile,
-            timeoutMs: toolTimeoutMs,
-            proxyRequest,
-            allowHostControl: opts?.allowHostControl,
-            sandboxBridgeUrl: opts?.sandboxBridgeUrl,
-            signal,
-          });
-        case "tabs":
-          return await executeTabsAction({
-            baseUrl,
-            profile,
-            timeoutMs: toolTimeoutMs,
-            proxyRequest,
-            targetId: bindingResult?.ok ? bindingResult.binding.targetId : undefined,
-            signal,
-          });
-        case "open": {
-          const targetUrl = readTargetUrlParam(params);
-          const label = normalizeOptionalString(params.label);
-          const opened = proxyRequest
-            ? await proxyRequest({
-                method: "POST",
-                path: "/tabs/open",
-                profile,
-                body: { url: targetUrl, ...(label ? { label } : {}) },
-                timeoutMs: toolTimeoutMs,
-              })
-            : await browserOpenTab(baseUrl, targetUrl, {
-                profile,
-                label,
-                timeoutMs: toolTimeoutMs,
-                signal,
-              });
-          const closeOpenedTab = async (targetId: string, openedProfile?: string) => {
-            if (nodeRoute && !proxyRequest?.isHostFallbackActive()) {
-              await nodeRoute.closeTarget({ targetId, profile: openedProfile });
-              return;
-            }
-            await browserCloseTab(baseUrl, targetId, {
-              profile: openedProfile,
-              timeoutMs: toolTimeoutMs,
-            });
-          };
-          await sessionTabs.trackOpened(opened, closeOpenedTab);
-          return formatBrowserExternalToolResult({
-            kind: "tabs",
-            payload: stripBrowserOpenInternalMetadata(opened),
-          });
-        }
-        case "focus": {
-          const targetId = readStringParam(params, "targetId", {
-            required: true,
-          });
-          const result = proxyRequest
-            ? await proxyRequest({
-                method: "POST",
-                path: "/tabs/focus",
-                profile,
-                body: { targetId },
-                timeoutMs: toolTimeoutMs,
-              })
-            : await browserFocusTab(baseUrl, targetId, {
-                profile,
-                timeoutMs: toolTimeoutMs,
-                signal,
-              });
-          sessionTabs.touch(
-            readStringValue((result as { targetId?: unknown }).targetId) ?? targetId,
-          );
-          return jsonResult(result);
-        }
-        case "close": {
-          const targetId = readStringParam(params, "targetId");
-          if (proxyRequest) {
-            const result = targetId
-              ? await proxyRequest({
-                  method: "DELETE",
-                  path: `/tabs/${encodeURIComponent(targetId)}`,
-                  profile,
-                  timeoutMs: toolTimeoutMs,
-                })
-              : await proxyRequest({
-                  method: "POST",
-                  path: "/act",
-                  profile,
-                  body: { kind: "close" },
-                  timeoutMs: toolTimeoutMs,
-                });
-            sessionTabs.untrack(
-              readStringValue((result as { targetId?: unknown }).targetId) ?? targetId,
-            );
-            return jsonResult(result);
-          }
-          const result = targetId
-            ? await browserCloseTab(baseUrl, targetId, {
-                profile,
-                timeoutMs: toolTimeoutMs,
-                signal,
-              })
-            : await browserAct(
-                baseUrl,
-                { kind: "close" },
-                {
-                  profile,
-                  timeoutMs: toolTimeoutMs,
-                  signal,
-                },
-              );
-          sessionTabs.untrack(readStringValue(result.targetId) ?? targetId);
-          return jsonResult(result);
-        }
-        case "snapshot":
-          return await executeSnapshotAction({
-            input: params,
-            baseUrl,
-            profile,
-            proxyRequest,
-            signal,
-            onTabActivity: sessionTabs.touch,
-          });
-        case "screenshot":
-          return await executeScreenshotAction({
-            input: params,
-            baseUrl,
-            profile,
-            requestedTimeoutMs,
-            proxyRequest,
-            signal,
-            onTabActivity: sessionTabs.touch,
-            opts,
-          });
-        case "navigate": {
-          const targetUrl = readTargetUrlParam(params);
-          const targetId = readStringParam(params, "targetId");
-          const timeoutMs =
-            requestedTimeoutMs === undefined
-              ? undefined
-              : resolveBrowserNavigationTimeoutMs(requestedTimeoutMs);
-          const result = proxyRequest
-            ? await proxyRequest({
-                method: "POST",
-                path: "/navigate",
-                profile,
-                body: {
-                  url: targetUrl,
-                  targetId,
-                  timeoutMs,
-                },
-                timeoutMs,
-              })
-            : await browserNavigate(baseUrl, {
-                url: targetUrl,
-                targetId,
-                timeoutMs,
-                profile,
-                signal,
-              });
-          const navigatedTargetId =
-            readStringValue((result as { targetId?: unknown }).targetId) ?? targetId;
-          sessionTabs.touch(navigatedTargetId);
-          const formatted = formatBrowserExternalToolResult({
-            kind: (result as { download?: unknown }).download ? "download" : "act",
-            payload: result,
-          });
-          // A navigation that resolved to a download leaves the document
-          // unchanged, so inline page state would describe the wrong thing.
-          if ((result as { download?: unknown }).download) {
-            return formatted;
-          }
-          return await appendNavigatedPageState({
-            result: formatted,
-            targetId: navigatedTargetId,
-            baseUrl,
-            profile,
-            proxyRequest,
-            signal,
-          });
-        }
-        case "console": {
-          const result = await executeConsoleAction({
-            input: params,
-            baseUrl,
-            profile,
-            proxyRequest,
-            signal,
-          });
-          const targetId = readStringParam(params, "targetId");
-          const canonicalTargetId = readStringValue(
-            (result.details as { targetId?: unknown } | undefined)?.targetId,
-          );
-          sessionTabs.touch(canonicalTargetId ?? targetId);
-          return result;
-        }
-        case "requests":
-        case "errors":
-        case "text":
-        case "emulate": {
-          const execute = {
-            requests: executeRequestsAction,
-            errors: executeErrorsAction,
-            text: executeTextAction,
-            emulate: executeEmulateAction,
-          }[action];
-          const result = await execute({ input: params, baseUrl, profile, proxyRequest, signal });
-          sessionTabs.touch(
-            readStringValue(asNullableRecord(result.details)?.targetId) ??
-              readStringValue(params.targetId),
-          );
-          return result;
-        }
-        case "pdf": {
-          const targetId = normalizeOptionalString(params.targetId);
-          const result = proxyRequest
-            ? ((await proxyRequest({
-                method: "POST",
-                path: "/pdf",
-                profile,
-                body: { targetId },
-              })) as Awaited<ReturnType<typeof browserPdfSave>>)
-            : await browserPdfSave(baseUrl, { targetId, profile, signal });
-          sessionTabs.touch(readStringValue(result.targetId) ?? targetId);
-          return {
-            content: [{ type: "text" as const, text: `FILE:${result.path}` }],
-            details: result,
-          };
-        }
-        case "download":
-        case "waitfordownload":
-          return await executeDownloadAction({
-            action,
-            input: params,
-            baseUrl,
-            profile,
-            proxyRequest,
-            signal,
-            onTabActivity: sessionTabs.touch,
-          });
-        case "upload": {
-          const paths = Array.isArray(params.paths) ? params.paths.map((p) => String(p)) : [];
-          if (paths.length === 0) {
-            throw new Error("paths required");
-          }
-          const resolvedResult = await resolveExistingUploadPaths({ requestedPaths: paths });
-          if (!resolvedResult.ok) {
-            throw new Error(resolvedResult.error);
-          }
-          const normalizedPaths = resolvedResult.paths;
-          const ref = readStringParam(params, "ref");
-          const inputRef = readStringParam(params, "inputRef");
-          const element = readStringParam(params, "element");
-          const { targetId, timeoutMs } = readOptionalTargetAndTimeout(params);
-          const request = {
-            paths: normalizedPaths,
-            ref,
-            inputRef,
-            element,
-            targetId,
-            timeoutMs,
-          };
-          return await executeTrackedTabRequest(
-            "/hooks/file-chooser",
-            request,
-            async () => await browserArmFileChooser(baseUrl, { ...request, profile, signal }),
-          );
-        }
-        case "dialog": {
-          const accept = Boolean(params.accept);
-          const promptText = readStringValue(params.promptText);
-          const dialogId = readStringValue(params.dialogId);
-          const { targetId, timeoutMs } = readOptionalTargetAndTimeout(params);
-          const request = { accept, promptText, dialogId, targetId, timeoutMs };
-          return await executeTrackedTabRequest(
-            "/hooks/dialog",
-            request,
-            async () => await browserArmDialog(baseUrl, { ...request, profile, signal }),
-          );
-        }
-        case "act": {
-          const request = readActRequestParam(params);
-          if (!request) {
-            throw new Error("request required");
-          }
-          if (!capabilities.actKinds.some((kind) => kind === request.kind)) {
-            throw new Error(
-              `browser act kind ${JSON.stringify(request.kind)} is unavailable for this run`,
-            );
-          }
-          return await executeActAction({
-            request,
-            baseUrl,
-            profile,
-            usesChromeMcp: isUserBrowserProfile,
-            proxyRequest,
-            signal,
-            onTabActivity: sessionTabs.touch,
-            onTabClose: sessionTabs.untrack,
-          });
-        }
-        default:
-          throw new Error(`Unknown action: ${action}`);
-      }
-    },
+    execute: createBrowserToolExecutor({
+      opts,
+      binding: bindingResult?.ok ? bindingResult.binding : undefined,
+      capabilities,
+    }),
   };
   return {
     ...tool,
