+++ src/plugins/registry-runtime.ts
@@ -23,24 +23,16 @@ import {
   isPluginRegistryRetired,
   revokePluginRecordLifecycleEpoch,
 } from "./registry-lifecycle.js";
+import { attachBrowserNodeDelegationRuntime } from "./registry-runtime-browser-delegation.js";
 import type { PluginRegistryState } from "./registry-state.js";
 import type { PluginRecord } from "./registry-types.js";
 import {
-  attachBrowserNodeDelegationResolver,
-  type BrowserNodeDelegationRuntime,
-} from "./runtime/browser-node-delegation.js";
-import {
   getGatewayContextResolver,
-  withPluginRuntimeGatewayRequestAuthority,
   withPluginRuntimePluginIdScope,
   withPluginRuntimePluginScope,
   withPluginRuntimeRegistryScope,
 } from "./runtime/gateway-request-scope.js";
 import type { PluginRuntime } from "./runtime/types.js";
-
-function isTrustedBrowserPluginRecord(record: PluginRecord | undefined): boolean {
-  return record?.origin === "bundled" || record?.trustedOfficialInstall === true;
-}
 
 export function createPluginRuntimeResolver(state: PluginRegistryState) {
   const { registry, registryParams } = state;
@@ -186,13 +178,13 @@ export function createPluginRuntimeResolver(state: PluginRegistryState) {
         | "openChannelIngressQueue"
         | "openChannelIngressDrain",
     ) => {
-      const record =
+      const currentRecord =
         pluginRuntimeRecordById.get(pluginId) ??
         registry.plugins.find((entry) => entry.id === pluginId);
-      if (record?.origin !== "bundled" && record?.trustedOfficialInstall !== true) {
+      if (currentRecord?.origin !== "bundled" && currentRecord?.trustedOfficialInstall !== true) {
         // Name the denied plugin and its origin so operators can replace the untrusted install.
         throw new Error(
-          `${methodName} is only available for trusted plugins in this release. Plugin "${pluginId}" loaded with origin "${record?.origin ?? "unknown"}"; reinstall it from its official npm package or ClawHub listing to enable trusted plugin state.`,
+          `${methodName} is only available for trusted plugins in this release. Plugin "${pluginId}" loaded with origin "${currentRecord?.origin ?? "unknown"}"; reinstall it from its official npm package or ClawHub listing to enable trusted plugin state.`,
         );
       }
     };
@@ -549,7 +541,8 @@ export function createPluginRuntimeResolver(state: PluginRegistryState) {
               !activePluginRuntimeRecords.has(channelOwnerRecord) ||
               isPluginRegistryRetired(registry) ||
               !registry.plugins.some(
-                (record) => record === channelOwnerRecord && record.status === "loaded",
+                (candidateRecord) =>
+                  candidateRecord === channelOwnerRecord && candidateRecord.status === "loaded",
               ) ||
               !registry.channels.some(
                 (channel) => channel.pluginId === pluginId && channel.plugin.id === messageChannel,
@@ -616,93 +609,15 @@ export function createPluginRuntimeResolver(state: PluginRegistryState) {
         } satisfies PluginRuntime["subagent"];
       },
     });
-    attachBrowserNodeDelegationResolver(runtime, (): BrowserNodeDelegationRuntime | undefined => {
-      const registration = registry.browserNodeDelegations.find((entry) =>
-        entry.delegation.consumerPluginIds.includes(pluginId),
-      );
-      if (!registration) {
-        return undefined;
-      }
-      const providerRecord = registration.providerRecord;
-      const isProviderRuntimeActive = () =>
-        pluginRuntimeRecordById.get(registration.pluginId) === providerRecord &&
-        activePluginRuntimeRecords.has(providerRecord) &&
-        isTrustedBrowserPluginRecord(providerRecord) &&
-        providerRecord.status === "loaded" &&
-        providerRecord.enabled &&
-        providerRecord.origin === registration.provider.origin &&
-        providerRecord.source === registration.provider.source &&
-        providerRecord.rootDir === registration.provider.rootDir &&
-        providerRecord.trustedOfficialInstall === registration.provider.trustedOfficialInstall;
-      if (!isProviderRuntimeActive()) {
-        return undefined;
-      }
-      if (
-        pluginRuntimeRecordById.get(pluginId) !== record ||
-        !isTrustedBrowserPluginRecord(record) ||
-        record.status !== "loaded" ||
-        !record.enabled
-      ) {
-        return undefined;
-      }
-      const consumerRecord = record;
-      let consumerEpoch = browserNodeDelegationEpochByRecord.get(consumerRecord);
-      const resolveConsumerEpoch = () => {
-        if (consumerEpoch) {
-          return consumerEpoch;
-        }
-        if (!isPluginRegistryActivated(registry) || isPluginRegistryRetired(registry)) {
-          return undefined;
-        }
-        if (
-          pluginRuntimeRecordById.get(pluginId) !== consumerRecord ||
-          !activePluginRuntimeRecords.has(consumerRecord) ||
-          consumerRecord.status !== "loaded" ||
-          !consumerRecord.enabled ||
-          !registry.plugins.some((entry) => entry === consumerRecord && entry.status === "loaded")
-        ) {
-          return undefined;
-        }
-        consumerEpoch = activatePluginRecordLifecycleEpoch(registry, consumerRecord);
-        if (consumerEpoch) {
-          browserNodeDelegationEpochByRecord.set(consumerRecord, consumerEpoch);
-        }
-        return consumerEpoch;
-      };
-      const isConsumerRuntimeActive = () => {
-        const epoch = resolveConsumerEpoch();
-        return Boolean(
-          epoch &&
-          pluginRuntimeRecordById.get(pluginId) === consumerRecord &&
-          activePluginRuntimeRecords.has(consumerRecord) &&
-          consumerRecord.status === "loaded" &&
-          consumerRecord.enabled &&
-          registry.plugins.some((entry) => entry === consumerRecord && entry.status === "loaded") &&
-          isPluginRecordLifecycleEpochActive(registry, consumerRecord, epoch),
-        );
-      };
-      const isBrowserNodeDelegationActive = () =>
-        registry.browserNodeDelegations.includes(registration) &&
-        isProviderRuntimeActive() &&
-        isConsumerRuntimeActive();
-      return {
-        request: async (params) => {
-          if (!registry.browserNodeDelegations.includes(registration)) {
-            throw new Error("Browser node delegation is no longer active.");
-          }
-          if (!isProviderRuntimeActive()) {
-            throw new Error("Browser node delegation provider lifecycle is no longer active.");
-          }
-          if (!isConsumerRuntimeActive()) {
-            throw new Error("Browser node delegation consumer lifecycle is no longer active.");
-          }
-          return await withPluginRuntimeGatewayRequestAuthority(isBrowserNodeDelegationActive, () =>
-            runWithPluginScope(() =>
-              registration.delegation.request({ ...params, consumerPluginId: pluginId }),
-            ),
-          );
-        },
-      };
+    attachBrowserNodeDelegationRuntime({
+      runtime,
+      registry,
+      pluginId,
+      record,
+      pluginRuntimeRecordById,
+      activePluginRuntimeRecords,
+      browserNodeDelegationEpochByRecord,
+      runWithPluginScope,
     });
     pluginRuntimeByRecord.set(record, runtime);
     return runtime;
